% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Plot the development of first names in France}
\usage{
draw_a_name(name)
}
\arguments{
\item{name}{}
}
\description{
Plot the development of first names in France
}
\examples{

}
