% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Draw the development of multiple first names in France since 1900}
\usage{
draw_names(name_vector)
}
\arguments{
\item{name_vector}{vector of names to be drawn}
}
\value{
The ggplot of names
}
\description{
Draw the development of multiple first names in France since 1900
}
\examples{
draw_names(c("Diane","Vincent"))
}
