% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names_dygraph.R
\name{draw_names_dygraph}
\alias{draw_names_dygraph}
\title{Draw the development of multiple first names in France since 1900 using dygraph}
\usage{
draw_names_dygraph(name_vector)
}
\arguments{
\item{name_vector}{vector of names to be drawn}
}
\value{
The plot of names
}
\description{
Draw the development of multiple first names in France since 1900 using dygraph
}
\examples{
draw_names(c("Diane","Vincent"))
}
