% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Draw the development of first names in France since 1900}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{a character}

\item{the_sex}{a character}
}
\value{
The ggplot of the name
}
\description{
Draw the development of first names in France since 1900
}
\examples{
draw_a_name("Vincent", "M")

}
